pyproject.toml
[tool.poetry]
name = "martini"
version = "0.1.0"
description = ""
authors = ["Davy Peter Braun <davy.braun@ginetta.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.11.1"
amqp = "^5.1.1"
anyio = "^3.7.0"
asgiref = "^3.7.2"
async-timeout = "^4.0.2"
asyncio-redis = "^0.16.0"
billiard = "^4.1.0"
broadcaster = "^0.2.0"
celery = "^5.3.0"
click-didyoumean = "^0.3.0"
click-plugins = "^1.1.1"
click-repl = "^0.2.0"
click = "^8.1.3"
colorama = "^0.4.6"
deprecated = "^1.2.14"
exceptiongroup = "^1.1.1"
fastapi = "^0.96.0"
flower = "^1.2.0"
greenlet = "^2.0.2"
h11 = "^0.14.0"
httptools = "^0.5.0"
humanize = "^4.6.0"
idna = "^3.4"
kombu = "^5.3.0"
mako = "^1.2.4"
markupsafe = "^2.1.3"
packaging = "^23.1"
prometheus-client = "^0.17.0"
prompt-toolkit = "^3.0.38"
pydantic = "^1.10.9"
python-dotenv = "^1.0.0"
pytz = "^2023.3"
pyyaml = "^6.0"
redis = "^4.5.5"
six = "^1.16.0"
sniffio = "^1.3.0"
sqlalchemy = "^2.0.15"
tornado = "^6.3.2"
typing-extensions = "^4.6.3"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
uvloop = "^0.17.0"
vine = "^5.0.0"
watchfiles = "^0.19.0"
wcwidth = "^0.2.6"
websockets = "^11.0.3"
wrapt = "^1.15.0"
langchain = "^0.0.194"
tiktoken = "^0.4.0"
pypdf = "^3.9.1"
pinecone-client = "^2.2.2"
openai = "^0.27.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
